backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
groq==0.11.0
httpx==0.27.2
httpcore==0.18.0
pydantic==2.9.2
pypdf==5.0.0
python-docx==1.1.2
chardet==5.2.0
